pyproject.toml
[tool.poetry]
name = "fast-lesson"
version = "0.1.0"
description = ""
authors = ["SabirovaOlya <sabirovaolya44@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.35"
sqladmin = "^0.19.0"
psycopg2-binary = "^2.9.9"
jwt = "^1.3.1"
python-decouple = "^3.8"
pyjwt = "^2.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
sqlalchemy-file = "^0.6.0"
fastapi-storages = "^0.3.0"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
